+++ src/LayoutParametersInjectionInterface.php
@@ -23,6 +23,8 @@ interface LayoutParametersInjectionInterface
      * ```
      *
      * @return array
+     *
+     * @psalm-return array<string, mixed>
      */
     public function getLayoutParameters(): array;
 }