+++ src/ViewRenderer.php
@@ -104,17 +104,20 @@ final class ViewRenderer implements ViewContextInterface
      * @param array $parameters The parameters (name-value pairs) that will be extracted
      * and made available in the view file.
      *
+     * @psalm-param array<string, mixed> $parameters
+     *
      * @return DataResponse The response instance.
      */
     public function render(string $view, array $parameters = []): DataResponse
     {
-        $contentParameters = $this->getContentParameters($parameters);
+        $contentParameters = $this->getContentParameters();
         $layoutParameters = $this->getLayoutParameters();
         $metaTags = $this->getMetaTags();
         $linkTags = $this->getLinkTags();
 
         return $this->responseFactory->createResponse(fn (): string => $this->renderProxy(
             $view,
+            $parameters,
             $contentParameters,
             $layoutParameters,
             $metaTags,
@@ -131,6 +134,8 @@ final class ViewRenderer implements ViewContextInterface
      * @param string $view The view name {@see WebView::render()}.
      * @param array $parameters The parameters (name-value pairs) that will be extracted
      * and made available in the view file.
+     *
+     * @psalm-param array<string, mixed> $parameters
      *
      * @return DataResponse The response instance.
      */
@@ -150,6 +155,8 @@ final class ViewRenderer implements ViewContextInterface
      * @param array $parameters The parameters (name-value pairs) that will be extracted
      * and made available in the view file.
      *
+     * @psalm-param array<string, mixed> $parameters
+     *
      * @throws RuntimeException If the view cannot be resolved.
      * @throws Throwable If an error occurred during rendering.
      * @throws ViewNotFoundException If the view file does not exist.
@@ -160,7 +167,8 @@ final class ViewRenderer implements ViewContextInterface
     {
         return $this->renderProxy(
             $view,
-            $this->getContentParameters($parameters),
+            $parameters,
+            $this->getContentParameters(),
             $this->getLayoutParameters(),
             $this->getMetaTags(),
             $this->getLinkTags(),
@@ -173,6 +181,8 @@ final class ViewRenderer implements ViewContextInterface
      * @param string $view The view name {@see WebView::render()}.
      * @param array $parameters The parameters (name-value pairs) that will be extracted
      * and made available in the view file.
+     *
+     * @psalm-param array<string, mixed> $parameters
      *
      * @throws RuntimeException If the view cannot be resolved.
      * @throws Throwable If an error occurred during rendering.
@@ -278,10 +288,15 @@ final class ViewRenderer implements ViewContextInterface
      * Renders a view as a string injecting parameters and tags into view context.
      *
      * @param string $view The view name {@see WebView::render()}.
-     * @param array $contentParameters The content parameters to inject.
-     * @param array $layoutParameters The layout parameters to inject.
+     * @param array $contentParameters The content parameters to render view.
+     * @param array $injectContentParameters The content parameters to inject.
+     * @param array $injectLayoutParameters The layout parameters to inject.
      * @param array $metaTags The meta tags to inject.
      * @param array $linkTags The link tags to inject.
+     *
+     * @psalm-param array<string, mixed> $contentParameters
+     * @psalm-param array<string, mixed> $injectContentParameters
+     * @psalm-param array<string, mixed> $injectLayoutParameters
      *
      * @throws RuntimeException If the view cannot be resolved.
      * @throws Throwable If an error occurred during rendering.
@@ -292,24 +307,30 @@ final class ViewRenderer implements ViewContextInterface
     private function renderProxy(
         string $view,
         array $contentParameters,
-        array $layoutParameters,
+        array $injectContentParameters,
+        array $injectLayoutParameters,
         array $metaTags,
         array $linkTags
     ): string {
-        $this->injectMetaTags($metaTags);
-        $this->injectLinkTags($linkTags);
+        $currentView = $this->view->withContext($this);
 
-        $this->view = $this->view->withContext($this);
-        $content = $this->view->render($view, $contentParameters);
+        $this->injectMetaTags($metaTags, $currentView);
+        $this->injectLinkTags($linkTags, $currentView);
+
+        $content = $currentView
+            ->withAddedCommonParameters($injectContentParameters)
+            ->render($view, $contentParameters);
 
         if ($this->layout === null) {
             return $content;
         }
 
-        $layoutParameters['content'] = $content;
-        $layout = $this->findLayoutFile($this->layout);
+        $layoutParameters = ['content' => $content] + $injectLayoutParameters;
+        $layout = $this->findLayoutFile($this->layout, $currentView);
 
-        return $this->view->renderFile($layout, $layoutParameters);
+        return $currentView
+            ->withAddedCommonParameters($layoutParameters)
+            ->renderFile($layout);
     }
 
     /**
@@ -321,8 +342,10 @@ final class ViewRenderer implements ViewContextInterface
      * @param array $renderParameters Parameters specified during rendering.
      *
      * @return array The injection content parameters merged with the parameters specified during rendering.
+     *
+     * @psalm-return array<string, mixed>
      */
-    private function getContentParameters(array $renderParameters): array
+    private function getContentParameters(): array
     {
         $parameters = [];
         foreach ($this->injections as $injection) {
@@ -330,13 +353,15 @@ final class ViewRenderer implements ViewContextInterface
                 $parameters = array_merge($parameters, $injection->getContentParameters());
             }
         }
-        return array_merge($parameters, $renderParameters);
+        return $parameters;
     }
 
     /**
      * Gets the merged injection layout parameters.
      *
      * @return array The merged injection layout parameters.
+     *
+     * @psalm-return array<string, mixed>
      */
     private function getLayoutParameters(): array
     {
@@ -389,13 +414,13 @@ final class ViewRenderer implements ViewContextInterface
      * @see WebView::registerMeta()
      * @see WebView::registerMetaTag()
      */
-    private function injectMetaTags(array $tags): void
+    private function injectMetaTags(array $tags, WebView $view): void
     {
         foreach ($tags as $key => $tag) {
             $key = is_string($key) ? $key : null;
 
             if (is_array($tag)) {
-                $this->view->registerMeta($tag, $key);
+                $view->registerMeta($tag, $key);
                 continue;
             }
 
@@ -410,7 +435,7 @@ final class ViewRenderer implements ViewContextInterface
                 );
             }
 
-            $this->view->registerMetaTag($tag, $key);
+            $view->registerMetaTag($tag, $key);
         }
     }
 
@@ -421,7 +446,7 @@ final class ViewRenderer implements ViewContextInterface
      *
      * @see WebView::registerLinkTag()
      */
-    private function injectLinkTags(array $tags): void
+    private function injectLinkTags(array $tags, WebView $view): void
     {
         foreach ($tags as $key => $tag) {
             if (is_array($tag)) {
@@ -457,7 +482,7 @@ final class ViewRenderer implements ViewContextInterface
                 }
             }
 
-            $this->view->registerLinkTag($tag, $position, is_string($key) ? $key : null);
+            $view->registerLinkTag($tag, $position, is_string($key) ? $key : null);
         }
     }
 
@@ -468,7 +493,7 @@ final class ViewRenderer implements ViewContextInterface
      *
      * @return string The path to the file with the file extension.
      */
-    private function findLayoutFile(string $file): string
+    private function findLayoutFile(string $file, WebView $view): string
     {
         $file = $this->aliases->get($file);
 
@@ -476,7 +501,7 @@ final class ViewRenderer implements ViewContextInterface
             return $file;
         }
 
-        return $file . '.' . $this->view->getDefaultExtension();
+        return $file . '.' . $view->getDefaultExtension();
     }
 
     /**