+++ tests/views/nested/nested-2.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @var \Yiisoft\View\WebView $this
+ * @var string $name
+ */
+
+echo ($label ?? 'nested-2') . ': ' . $name . '.';